tests: Adding test cases for images without tags
